EOS-9804-improvement: Split remote command execution API to separate python file. Add default parameters for cmd_run API
